+++ test-app/src/types/openapi/openapi.d.ts
@@ -183,7 +183,7 @@ export interface paths {
                         uuid?: string;
                         /** Format: decimal */
                         volume?: number | null;
-                        password?: (string | null) | (number | null) | (Record<string, never> | null);
+                        password?: string | null;
                         openapiVirtualSpecTest?: string | null;
                         openapiVirtualSpecTest2?: string[];
                     };
@@ -304,7 +304,7 @@ export interface paths {
                         uuid?: string;
                         /** Format: decimal */
                         volume?: number | null;
-                        password?: (string | null) | (number | null) | (Record<string, never> | null);
+                        password?: string | null;
                         openapiVirtualSpecTest?: string | null;
                         openapiVirtualSpecTest2?: string[];
                     };
@@ -509,6 +509,126 @@ export interface paths {
                     };
                     content: {
                         "application/json": (components["schemas"]["BalloonLatex"] | components["schemas"]["BalloonMylar"] | components["schemas"]["Pet"] | components["schemas"]["ViewModelsMyViewModel"])[];
+                    };
+                };
+                400: components["responses"]["BadRequest"];
+                401: components["responses"]["Unauthorized"];
+                403: components["responses"]["Forbidden"];
+                404: components["responses"]["NotFound"];
+                409: components["responses"]["Conflict"];
+                418: components["responses"]["CustomResponse"];
+                422: components["responses"]["ValidationErrors"];
+                490: components["responses"]["CustomResponse"];
+                500: components["responses"]["InternalServerError"];
+            };
+        };
+        put?: never;
+        post?: never;
+        delete?: never;
+        options?: never;
+        head?: never;
+        patch?: never;
+        trace?: never;
+    };
+    "/balloons/paginated": {
+        parameters: {
+            query?: {
+                /** @description Page number */
+                page?: string;
+            };
+            header?: {
+                /** @description custom header */
+                "custom-header"?: string;
+            };
+            path?: never;
+            cookie?: never;
+        };
+        get: {
+            parameters: {
+                query?: {
+                    /** @description Page number */
+                    page?: string;
+                };
+                header?: {
+                    /** @description custom header */
+                    "custom-header"?: string;
+                };
+                path?: never;
+                cookie?: never;
+            };
+            requestBody?: never;
+            responses: {
+                /** @description Success */
+                200: {
+                    headers: {
+                        [name: string]: unknown;
+                    };
+                    content: {
+                        "application/json": {
+                            recordCount: number;
+                            pageCount: number;
+                            currentPage: number;
+                            results: (components["schemas"]["BalloonLatex"] | components["schemas"]["BalloonMylar"])[];
+                        };
+                    };
+                };
+                400: components["responses"]["BadRequest"];
+                401: components["responses"]["Unauthorized"];
+                403: components["responses"]["Forbidden"];
+                404: components["responses"]["NotFound"];
+                409: components["responses"]["Conflict"];
+                418: components["responses"]["CustomResponse"];
+                422: components["responses"]["ValidationErrors"];
+                490: components["responses"]["CustomResponse"];
+                500: components["responses"]["InternalServerError"];
+            };
+        };
+        put?: never;
+        post?: never;
+        delete?: never;
+        options?: never;
+        head?: never;
+        patch?: never;
+        trace?: never;
+    };
+    "/balloons/scroll-paginated": {
+        parameters: {
+            query?: {
+                /** @description Fast pagination cursor */
+                cursor?: string | null;
+            };
+            header?: {
+                /** @description custom header */
+                "custom-header"?: string;
+            };
+            path?: never;
+            cookie?: never;
+        };
+        get: {
+            parameters: {
+                query?: {
+                    /** @description Fast pagination cursor */
+                    cursor?: string | null;
+                };
+                header?: {
+                    /** @description custom header */
+                    "custom-header"?: string;
+                };
+                path?: never;
+                cookie?: never;
+            };
+            requestBody?: never;
+            responses: {
+                /** @description Success */
+                200: {
+                    headers: {
+                        [name: string]: unknown;
+                    };
+                    content: {
+                        "application/json": {
+                            cursor: string | null;
+                            results: (components["schemas"]["BalloonLatex"] | components["schemas"]["BalloonMylar"])[];
+                        };
                     };
                 };
                 400: components["responses"]["BadRequest"];
@@ -2325,7 +2445,7 @@ export interface paths {
                         uuid?: string;
                         /** Format: decimal */
                         volume?: number | null;
-                        password?: (string | null) | (number | null) | (Record<string, never> | null);
+                        password?: string | null;
                         openapiVirtualSpecTest?: string | null;
                         openapiVirtualSpecTest2?: string[];
                     };
@@ -2521,7 +2641,7 @@ export interface paths {
                         uuid?: string;
                         /** Format: decimal */
                         volume?: number | null;
-                        password?: (string | null) | (number | null) | (Record<string, never> | null);
+                        password?: string | null;
                         openapiVirtualSpecTest?: string | null;
                         openapiVirtualSpecTest2?: string[];
                     };
@@ -2799,7 +2919,7 @@ export interface paths {
                         uuid?: string;
                         /** Format: decimal */
                         volume?: number | null;
-                        password?: (string | null) | (number | null) | (Record<string, never> | null);
+                        password?: string | null;
                         openapiVirtualSpecTest?: string | null;
                         openapiVirtualSpecTest2?: string[];
                         /** @description Page number */
@@ -2985,7 +3105,7 @@ export interface paths {
                         uuid?: string;
                         /** Format: decimal */
                         volume?: number | null;
-                        password?: (string | null) | (number | null) | (Record<string, never> | null);
+                        password?: string | null;
                         openapiVirtualSpecTest?: string | null;
                         openapiVirtualSpecTest2?: string[];
                         /** @description Fast pagination cursor */