+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.14.0
+
+- generated code uses absolute imports
+- account for change in Dream in which `Virtual` decorator requires OpenAPI shape
+- throw a more detailed error when DreamSerializer.attribute used to render a non-database, non-Virtual decorated property
+- fix OpenAPI generated when paginating/scrollPaginating an STI base model
+
 ## 1.13.0
 
 - bump to Dream 1.12.0, which changes DateTime and CalendarDate to throw an exception rather than allowing invalid datetimes/dates